revert extractartifacts changes
