+++ plugins/codeamp/graphql/projectextension_mutation.go
@@ -10,7 +10,6 @@ import (
 	"github.com/codeamp/circuit/plugins"
 	"github.com/codeamp/circuit/plugins/codeamp/auth"
 	db_resolver "github.com/codeamp/circuit/plugins/codeamp/db"
-	"github.com/codeamp/circuit/plugins/codeamp/helpers"
 	"github.com/codeamp/circuit/plugins/codeamp/model"
 	log "github.com/codeamp/logger"
 	"github.com/codeamp/transistor"
@@ -82,7 +81,7 @@ func (r *ProjectExtensionResolverMutation) CreateProjectExtension(ctx context.Co
 
 		r.DB.Save(&projectExtension)
 
-		artifacts, err := helpers.ExtractArtifacts(projectExtension, extension, r.DB)
+		artifacts, err := ExtractArtifacts(projectExtension, extension, r.DB)
 		if err != nil {
 			log.Error(err.Error())
 			return nil, err
@@ -155,7 +154,7 @@ func (r *ProjectExtensionResolverMutation) UpdateProjectExtension(args *struct{ 
 
 	r.DB.Save(&projectExtension)
 
-	artifacts, err := helpers.ExtractArtifacts(projectExtension, extension, r.DB)
+	artifacts, err := ExtractArtifacts(projectExtension, extension, r.DB)
 	if err != nil {
 		log.Error(err.Error())
 		return nil, err
@@ -220,7 +219,7 @@ func (r *ProjectExtensionResolverMutation) DeleteProjectExtension(args *struct{ 
 
 	r.DB.Delete(&projectExtension)
 
-	artifacts, err := helpers.ExtractArtifacts(projectExtension, extension, r.DB)
+	artifacts, err := ExtractArtifacts(projectExtension, extension, r.DB)
 	if err != nil {
 		log.Error(err.Error())
 		return nil, err
@@ -258,7 +257,7 @@ func (r *ProjectExtensionResolverMutation) HandleExtensionRoute53(args *struct{ 
 		return fmt.Errorf("Could not unmarshal custom config")
 	}
 
-	artifacts, err := helpers.ExtractArtifacts(*projectExtension, extension, r.DB)
+	artifacts, err := ExtractArtifacts(*projectExtension, extension, r.DB)
 	if err != nil {
 		log.Error(err)
 		return err
@@ -286,7 +285,7 @@ func (r *ProjectExtensionResolverMutation) HandleExtensionRoute53(args *struct{ 
 				return err
 			}
 
-			tmpExtensionArtifacts, err := helpers.ExtractArtifacts(existingProjectExtension, tmpExtension, r.DB)
+			tmpExtensionArtifacts, err := ExtractArtifacts(existingProjectExtension, tmpExtension, r.DB)
 			if err != nil {
 				return err
 			}
