+++ plugins/codeamp/release_handler.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/codeamp/circuit/plugins"
 	graphql_resolver "github.com/codeamp/circuit/plugins/codeamp/graphql"
-	"github.com/codeamp/circuit/plugins/codeamp/helpers"
 	"github.com/codeamp/circuit/plugins/codeamp/model"
 	log "github.com/codeamp/logger"
 	"github.com/codeamp/transistor"
@@ -84,7 +83,7 @@ func (x *CodeAmp) ReleaseEventHandler(e transistor.Event) error {
 				}
 
 				if needsExtract {
-					artifacts, err = helpers.ExtractArtifacts(projectExtension, extension, x.DB)
+					artifacts, err = ExtractArtifacts(projectExtension, extension, x.DB)
 					if err != nil {
 						log.Error(err.Error())
 						return nil
