+++ plugins/codeamp/resourceconfig/projectextension.go
@@ -1,9 +1,12 @@
 package resourceconfig
 
 import (
+	"encoding/json"
 	"fmt"
 
-	"github.com/codeamp/circuit/plugins/codeamp/helpers"
+	log "github.com/codeamp/logger"
+	"github.com/codeamp/transistor"
+
 	"github.com/codeamp/circuit/plugins/codeamp/model"
 	"github.com/jinzhu/gorm"
 	yaml "gopkg.in/yaml.v2"
@@ -49,12 +52,32 @@ func (p *ProjectExtensionConfig) Export() (*ProjectExtension, error) {
 		return nil, err
 	}
 
-	artifacts, err := helpers.ExtractArtifacts(*p.projectExtension, extension, p.db)
+	configArtifacts := []transistor.Artifact{}
+
+	unmarshaledProjectExtensionConfig := []transistor.Artifact{}
+	err := json.Unmarshal(p.projectExtension.Config.RawMessage, &unmarshaledProjectExtensionConfig)
 	if err != nil {
 		return nil, err
 	}
 
-	artifactsBytes, err := yaml.Marshal(artifacts)
+	for _, artifact := range unmarshaledProjectExtensionConfig {
+		secretValue := model.SecretValue{}
+		// config artifact value is a reference to the actual secret object
+		secretID := artifact.Value
+		if p.db.Where("secret_id = ?", secretID).Order("created_at desc").First(&secretValue).RecordNotFound() {
+			log.InfoWithFields("secret value not found", log.Fields{
+				"secret_id": secretID,
+			})
+		}
+
+		configArtifact := transistor.Artifact{
+			Key:   artifact.Key,
+			Value: secretValue.Value,
+		}
+		configArtifacts = append(configArtifacts, configArtifact)
+	}
+
+	artifactsBytes, err := yaml.Marshal(configArtifacts)
 	if err != nil {
 		return nil, err
 	}
