+++ plugins/codeamp/workflowreleaseextension_handler.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/codeamp/circuit/plugins"
 	graphql_resolver "github.com/codeamp/circuit/plugins/codeamp/graphql"
-	"github.com/codeamp/circuit/plugins/codeamp/helpers"
 	"github.com/codeamp/circuit/plugins/codeamp/model"
 	log "github.com/codeamp/logger"
 	"github.com/codeamp/transistor"
@@ -222,7 +221,7 @@ func (x *CodeAmp) WorkflowReleaseExtensionsCompleted(release *model.Release) {
 				return
 			}
 
-			projectExtensionArtifacts, err := helpers.ExtractArtifacts(projectExtension, extension, x.DB)
+			projectExtensionArtifacts, err := ExtractArtifacts(projectExtension, extension, x.DB)
 			if err != nil {
 				log.Error(err.Error())
 			}
